Define script@nonce and style@nonce attributes
The definitions here are little more than placeholders. We need to
improve the description in CSP such that HTML can hook into reasonable
definitions of the behavior that nonces require when executing the
"prepare a script" and "update a style block" algorithms. For the
moment, defining the attributes and pointing out to the CSP spec (and
to w3c/webappsec#463) is probably Good Enough(tm).