Note elements which cannot be shadow hosts
Fixes #85 by creating a new category of elements which cannot be shadow
hosts, and porting over the list from the shadow DOM specification.

See also https://github.com/w3c/webcomponents/issues/110#issuecomment-137114260.

This is also related to #86, as currently the list corresponds to
elements whose rendering is defined with a binding property, but we
want to remove the binding stuff anyway.