Bypass appcache in favor of a service worker
When both a service worker registration and an application cache are
available for the navigation of resources, if the navigation is under
the scope of a service worker, completely bypass the appcache regardless
of whether the navigation uses the service worker.

Fixes https://www.w3.org/Bugs/Public/show_bug.cgi?id=27482.

PR #184.