Fix #194: remove PortCollection in favor of advice
The design of PortCollection was such that it exposed the details of
garbage collection. Since that is unacceptable to implementations it
has not been implemented and this commit removes the feature.
Unfortunately this means developers will have to do manual memory
management for ports. To address that this commit gives some advice on
how to do that.