Make constraint validation algorithm for minlength attribute match reality
Per http://caniuse.com/#feat=input-minlength ,
only Blink currently implements minlength, and it implements it similar to
maxlength, which means it requires the user to edit the value in order for it
to be considered "too short".
So this commit copies the dirtiness and "last edited by the user" conditions
from maxlength's constraint validation algorithm
into minlength's constraint validation algorithm.

Closes #174.