Fix #247: make SVG titles first child, not last
The SVG spec says that the title should be the first child, and that
some UAs might only respect such children, although non-first children
are also valid.  The only UA that seems to implement setting
document.title in SVG documents, IE, inserts as the first child and not
the last.