Replace 2dcontext's currentTransform with get/setTransform
After discussion in #264 we concluded that it makes more sense for the canvas
2dcontext object to expose currentTransform as set and get methods rather than
an attribute. This avoids the complications of having a "live" object.

However, since we already have a setTransform() method, the solution was
to replace currentTransform with a getTransform() getter, and overload
setTransform() to also accept a DOMMatrixInit object.

Also clean up the wording around the term "current transformation matrix".