Add a 'noopener' <link rel> keyword and window feature
Based on [Charlie Reis' proposal][1], this patch adds a mechanism
by which web developers can open a new window without granting that
window access to the opener browsing context. This is possible today
via 'rel=noopener target=_blank', but that is both obscure and poorly
explanatory. This patch adds a 'noopener' keyword to both hyperlinks
and 'window.open' which clarifies the behavior and makes it available
without side effects.

[1]: https://wiki.whatwg.org/wiki/Links_to_Unrelated_Browsing_Contexts