Refactor browsing context creation
This reorganizes the section on browsing context creation so that creating a browsing context is now spelled out as a series of steps, and is more explicit about the association and creation of Window, Document, and WindowProxy objects.

This is part of the plan noted in https://www.w3.org/Bugs/Public/show_bug.cgi?id=25981 for integration with the new ECMAScript initialization lifecycle.