Fix duplicate dfn/subdfn for canvas's setTransform
Until recently we had only one way to call the canvas setTransform() method:
with a list of doubles. Then bb00977fe0f4960f929bb9fa136be2003383d551
overloaded the method to also accept a DOMMatrixInit object. But the change
causes a spec build error because the markup for it in the spec source causes
duplicate IDs in the generated spec output. This change de-duplicates.