Make ImageBitmap transferable
ImageBitmap objects can already be sent using postMessage(), but this
requires a duplication of the bitmap data in memory, which is
inefficient and often unnecessary. This change is a pre-requisite
for the OffscreenCanvas proposal, which relies on ImageBitmaps as
an efficient means of transferring bitmap data to and from Workers.
See: https://wiki.whatwg.org/wiki/OffscreenCanvas