Report errors to developer consoles, not users
In #224 we had a brief discussion which concluded that these statements about error reporting really referred to a developer console. Alternately, maybe they are holdovers from the early days of the web when browsers did in fact show popups with every JavaScript error. Regardless, this uniformly updates all script error reporting to be to a developer console, instead of to the user.