Fix example that used uncool URLs
The example for <input type="url"> coupled with a <datalist> referenced
several URLs that are uncool (i.e. have changed over time to become
broken). Unfortunately, some of them were baked into an image giving a
sample rendering of the control. As such, we replace the example with a
new one that demonstrates the same mechanisms, but uses new URLs, and we
replace the corresponding Windows 9x-era PNG screenshot with a more
modern-looking SVG rendering.

Closes #565.