Clean up SVGs, relying on only viewBox, and no space before slash
The viewBox width of content-venn.svg was 1024 but the used width was
1000, which merely adds some padding as aspect ratio is preserved.
Fix the viewBox to match the content, which is in fact 1000 units wide.

parsing-model-overview.svg had no viewBox, which means that it had no
intrinsic aspect ratio and wasn't scalable at all; if the size of the
embedding <object> were any different, the graphics would simply have
clipped. Use the <object>'s size as the viewBox size to get the same
rendering as before but without this hidden oddity.

PR: https://github.com/whatwg/html/pull/579