Add module workers
Closes #550. The Worker and SharedWorker constructors now get an options
object, which can be specified as { type: "module", credentials }, where
credentials specifies the credentials mode used for the initial fetch
and for subsequent imports. (This only applies to module workers; in
the future we could extend it to classic workers if we wished.)

importScripts will always throw a TypeError in modules, per discussion
in the pull request.

The module fetching and execution machinery is entirely reused from that
for <script type="module">. This also refactors the machinery for
fetching a classic worker script out into its own sub-algorithm of the
"Fetching scripts" section.