Remove entity-related files
This moves all entity-related files to https://github.com/whatwg/html-build. It is part of a larger changed discussed in https://github.com/whatwg/html-build/issues/62 to make building the HTML standard easier and require dedicated issues for changing the entities and CLDR data that are used by the HTML standard.