Use "credentials mode" instead of "CORS setting" for modules
This is a purely editorial update. We have essentially two types
representing the credentials mode. Previously we used the CORS setting
type up until the last minute, where we converted it into a credentials
mode. This ends up being inconvenient for worker-modules, so instead we
now convert from CORS setting to credentials mode up front and use that
throughout.