Allow pingback/prefetch/stylesheet links in body
Some external links augment the current document and should not be
restricted to metadata content. For example, the server may want to emit
a prefetch link within body content based on the generated content (well
after the head section was flushed).

Similarly, the position of a rel=stylesheet link in body can act as an
optimization and signal to the user agent that DOM content above it may
be painted - see [1]. However, note that this update does not define or
change how or when CSS is applied - that's a separate discussion.

All current browsers allow and process external resource links in body,
this update reflects what's implemented and being used by developers.

[1] https://www.w3.org/Bugs/Public/show_bug.cgi?id=27303#c37

PR: https://github.com/whatwg/html/pull/616