Use new "document's browsing context" concept everywhere
a398377bbc8483e27c4fc478a097ef72f3057593 formalized the concept of
"document's browsing context", but left most places that reference it as
simple links to "browsing context". This commit is the result of
auditing all "browsing context" references and pointing those that are
specific to a document to the new "document's browsing context" concept.