Allow truncation of alert/confirm/prompt text
As discussed recently on blink-dev, these dialogs are often used for
scams and such. At least Chrome has already applied the mitigation
strategy of limiting the length of the author-supplied text to 2K
characters (and @avidrissman proposes that Chrome limit it further). We
should call out in the spec that this is explicitly allowed, like we
already do for the beforeunload text.

https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/cFtsKXKd4F0
has some background.