Fix #559: require focus movement on scrolling to a fragment ID
This makes "scroll to the fragment identifier" require moving the
sequential focus navigation starting point, instead of letting that be
optional.

An example of this in action is clicking on a link in a table of
contents; what happens after doing so, and then pressing TAB?
Previously user agents were allowed to jump back to the list of links
in the table of contents, instead of starting the tab navigation from
the target of the link.

PR: https://github.com/whatwg/html/pull/726