Indicate that ImageData object creation rethrows RangeError exceptions
This is a normative change that requires the ImageData constructor and
the createImageData and getImageData methods to rethrow exceptions from
from the creation of the ImageData object. The motivation is to improve
the stability of web browsers by avoiding out-of-memory crashes when
the creation of ImageData objects fails. The ECMAScript specification
states that failure to allocate an ArrayBuffer results in a RangeError
exception being thrown, which should now be rethrown.

Reference:
http://ecma-international.org/ecma-262/6.0/#sec-createbytedatablock