Fix #197: indicate that ImageData creation rethrows RangeErrors
This is a normative change that requires the ImageData constructor and
the createImageData and getImageData methods to rethrow exceptions from
from the creation of the ImageData object. The motivation is to improve
the stability of web browsers by avoiding out-of-memory crashes when
the creation of ImageData objects fails. The JavaScript specification
states that failure to allocate an ArrayBuffer results in a RangeError
exception being thrown, which should now be rethrown. See:
https://tc39.github.io/ecma262/#sec-createbytedatablock.

PR: https://github.com/whatwg/html/pull/520