Sort out the handling of seconds in "parse a time component"
The nonsensical "let the collected string be second" was spotted and
fixed by @chaals in the W3C fork: https://github.com/w3c/html/pull/115

The usual style is to "let" a variable when introducing it and to "set"
it when updating it, so do that instead. The new structure ought to be
closer to how it would actually be implemented, but no less clear.