The DOM Standard defines the class, id, and slot attributes
No longer give the appearance that the HTML Standard defines the user
agent requirements around these attributes. Instead defer to the DOM
Standard, while still defining conformance requirements as relevant
to HTML elements.

This also lists the slot "superglobal" attribute for the first time
as it was only recently added to the DOM Standard.

Fixes #895 and is also a tiny step toward integrating Shadow DOM.