Clarify that createImageBitmap fails with an SVG blob.
This change makes it explicit that creating an ImageBitmap from a Blob
only supports *bitmap* file formats. This behaviour is consistent with
creating an ImageBitmap from an img element, which explicitly does not
support vector graphics.

Note: Supporting vector graphics would require defining a sizing
algorithm to determine the scale of image, or to have an alternate API
that has additional size parameters.