Remove some requirements on how to implement media element controls
The requirements on a play button make sense if there are separate
rewind and fastforward buttons where the play button goes back to the
normal playback rate, but that should not be required. As actually
implemented in browsers right now, the play button doesn't reset
playback rate.

Also drop the second half of the paragraph, which is redundant with the
preceding paragraph.

Fixes #956.