Wire up HostEnsureCanCompileStrings to CSP
CSP defines an algorithm we can use to define
HostEnsureCanCompileStrings, which throws an EvalError if string
compilation is disallowed.

We define HostEnsureCanCompileStrings here, delegating to CSP's
EnsureCSPDoesNotBlockStringCompilation, because in the future other
specs might also want to limit eval and setTimeout. It also centralizes
all the HostWhatever abstract operations in one place, which is nice.

Fixes #271.