Specify service worker registration selection
This PR defines an active worker for a Document and a
WorkerGlobalScope and the corresponding steps that set the values when
they are created. It also defines an active worker algorithm in the
environment settings object that returns the stored active worker.

This commit also makes the active worker of a document, whose address
is local and who has its creator (a parent or an opener), initialized
to the creator's document's active worker. The relevant steps are
added both in the creating a new browsing context algorithm and in the
initializing a new document algorithms. (Fixes #321)

R= @annevk.