Fix races between tasks resolving/rejecting play promises and them being added
For example, if pause() is called, it creates a task to reject all
promises even if there is none around. If play() is called just after,
this task would reject the newly created promise.

This is fixing the issue by creating tasks rejecting/resolving a given
list of promises, copied for the task.

Closes #996.