Use "create an element" whenever we create an element
Fixes #1065 by explicitly using the new "create an element" algorithm
from DOM whenever we create an element, and explicitly passing in the
appropriate node document and namespace while doing so.

Also updates the one existing call to "create an element" to account for
the new parameter ordering as of https://github.com/whatwg/dom/pull/231,
and slightly modernizes the algorithm style for several table methods.