Fix #1057: createImageBitmap to throw when img has 0 size
Since the spec has no provisions for handling ImageBitmaps that
have no bitmap data, it is important to explicitly disallow
the creation of ImageBitmap with a width or height of 0. With this
change, createImageBitmap will throw an exception when a source
image has a dimension of 0 *and* no crop rect is specified.