Formalize custom element reactions
This executes a large portion of the plan laid out in https://github.com/w3c/webcomponents/issues/186, by introducing a new [CEReactions] IDL extended attribute which can be used to indicate that new steps should be inserted related to custom element reactions. It replaces the previous handwaving of "every operation and attribute" with an explicit annotation on every relevant operation, attribute, setter, and deleter.

There remains work to do to annotate other specifications across the web platform with this attribute (notably DOM), as well as deal with editing operations. But all such updates will depend on the definition of [CEReactions] added here.