Generalize inserted/removed into/from a document algorithms
This updates the "inserted into a document" and "removed from a document" algorithms to work on all nodes, not just elements. It also redefines them in terms of the DOM Standard's insertion/removing steps for increased precision.

This sets the stage for various shadow DOM-related changes, such as the work on <script> elements necessary for #762.