Regression: do not use "get an output encoding" for decoding
In 6a31c26cf12e39dab1a488e75dd56c03d6786d39 I applied “get an output encoding” a bit enthusiastically, which gives the wrong result for the replacement encoding. It is only to be used when encoding (e.g., <form> submission and URLs), not decoding.

This fixes the regression part of #1077.