Link relations: add dns-prefetch/preconnect/prerender; update prefetch
This adds dns-prefetch, preconnect, and prerender to the list of link
types, defines that they are allowed only on <link>, and adds them to
the relevant supported tokens lists. The processing model defers to the
Resource Hints specification.

Similarly, this commit also updates the existing prefetch link type to
only be allowed on <link> (per browser reality as shown in
https://github.com/whatwg/html/pull/1141#issuecomment-218060752), and
defers its processing model to the Resource Hints spec.

Closes #1109.