currentScript needs to point to removed script elements
To avoid leaking shadow trees we added an “in a document” check in
caf203fd4d9a5047d1ff2edad2290d53e21368ba. This turned out to not be
web compatible. So instead, we check whether the script element is in
a shadow tree. Ideally that is novel enough to not cause any
compatibility issues.

Fixes #1161.