Integrate with the Referrer Policy spec, part 1 of n
This ports over the content and IDL attribute definitions, as well as
the <meta name="referrer"> definition, from the Referrer Policy
specification. In the process, it clarifies several things that document
did not:

- Conformance requirements for authors
- The missing and invalid value defaults are the empty string referrer
  policy
- <meta name="referrer"> only works in the document tree (not in a
  shadow tree)

There is more work to do; see especially:

- https://github.com/w3c/webappsec-referrer-policy/issues/40
- https://github.com/w3c/webappsec-referrer-policy/issues/30

This closes https://github.com/w3c/webappsec-referrer-policy/issues/27.

While working on the <meta> definition, this tidies up the theme-color
definition in similar ways.