Fix customElements.whenDefined() behavior
The previous text said to not create a strong reference to the promise,
which was incorrect since it means that code like
customElements.whenDefined("x-foo").then(f) might never call f, since
nothing is keeping the promise alive. So, that needs to be removed.

Once that is removed, keeping a separate promise around for each call to
whenDefined() becomes more expensive. So, instead we create a maximum of
one promise, and return it every time.