Use Travis CI to test building the source
This is a simple patch that allows us to test spec changes in branches
and pull requests by running the build on Travis CI. Currently it relies
on the build.whatwg.org build-server fallback path of html-build; we
hope to reduce this external dependency over time.

Note that this does not perform any deployment; that is still handled by
a post-commit hook. The purpose of this is simply to give feedback on
whether a commit successfully builds, and not to use Travis CI to deploy
from master.

Fixes #1250. Further discussion of goals for improving the CI is in
https://github.com/whatwg/html-build/issues/46.