Make a.ping and tableCell.headers strings, not DOMTokenLists
Fixes #1026. All current browser implement these as strings, not token
lists; the spec was aspirational in this regard, albeit with vendor
support. However, as discussed in
https://github.com/whatwg/html/issues/1026#issuecomment-219121243,
changing a.ping is not likely to be web-compatible, and changing
tableCell.headers becomes much less interesting if we're not going to be
able to uniformly say that DOMTokenLists are used for token-containing
attributes.