Clean up named and indexed property algorithm definitions
Previously, the spec had definitions of "determine the value of a named
property", "set", and "delete an existing named property", which was
weird because they only applied to DOMStringMap. Now we properly
cross-reference Web IDL for those algorithms, and instead the
getter/setter/deleter keywords link to the paragraphs in which those
algorithms are defined for each class that implements them. This
involved canvasing all places they were used and making sure they all
work the same way, which was not previously the case.

Fixes #922.