Merge DOMStringMap into the data-* section
Previously DOMStringMap had an overly-generic architecture, meant to
serve both the microdata API and data-* attributes. Since the microdata
API is now removed, and since this is a horrible interface which nobody
should ever use in the future, we can safely un-genericize it and merge
it into the section on data-* attributes.

In the process, made some minor updates to the algorithms to increase
precision, and cross-linked the XML Name production in a few other
places in the document.