Use USVString for all URLs
This finds all instances where URLs are part of an IDL interface,
including where they are reflecting content attributes, and changes the
type from DOMString to USVString. This has already been done in some
places in a haphazard way, but this commit finishes the process.

This allows us to clarify the rules for reflecting USVString attributes,
and remove the clause that allows DOMString attributes to reflect URLs.

Fixes #1254. Fixes #1266.