location.hash = "" should result in a fragment
Although some browsers currently do not add a fragment if hash is set
to the empty string, they won’t remove a fragment either when hash is
set to the empty string. Therefore aligning with Firefox makes more
sense, which consistently gives a fragment.

This makes the location API slightly inconsistent from the point of
<a>, <area>, and URL, but compatibility has to win.

Fixes #1177.