Regression: ensure SharedWorkerGlobalScope does not leak across origins
The SharedWorker always needs to perform a same-origin check since
although with modules the URL itself can be cross-origin, the origin of
the global of the WorkerGlobalScope needs to be same-origin.

This also cleans up “set up a worker environment settings object” with
a more explicit way of referencing the needed environment settings
object.

This regressed through #339 (though that change also had to be made for
module workers).

Fixes #1321.