Reduce the number of Window objects exposed on the Window object
In particular, when a child browsing context is cross-origin from the
perspective of the document it's in, only include it as a property on
the Window object when its browsing context name matches that of the
name content attribute value it is included through.

This also cleans up the definition of browsing context name by
aligning it with how it's used. Namely that it's the empty string by
default rather than unset. It also makes it clearer that "child
browsing context name property set" belongs to a Window object and
how the referenced "active document" is obtained from that Window
object.

Fixes #257.