Fix inaccurate stack-popping in custom element upgrades
Fixes #1238. Usually, the popped stack entry is not the element, but
instead an already-constructed marker. This makes the stack-popping
generic, and also adds a note explaining exactly what could be going on
here.

Also clarifies the cases in which the already constructed marker
matters. In addition to calling super() multiple times, this error case
can be triggered by creating another instance of the class being
constructed. This adds examples of both possibilities.

Finally, fixes a typo in the HTMLElement() constructor which used the
variable names "instance" and "element" interchangeably, and fixes a
broken cross-reference to "upgrade an element".