Editorial: "allowed to show a popup" → "triggered by user activation"
The "allowed to show a popup" algorithm can be more generally useful
than just in popup blocking. It should get a more general name so that
it can be reused more broadly.

This commit also moves it into the "Activation" section, instead of the
"Browsing context names" section, as the former makes a lot more sense.