Make navigate's resource handling more explicit
This in general integrates navigation with Fetch better, reusing Fetch's
infrastructure for network errors and for 401 handling. This also
removes the speculative 401 features that seem unlikely to ever
materialize.

This definitely requires more work to actually talk about responses
instead of resources, and to more clearly tie it to the "resource"
variable from earlier in the navigate algorithm, but since all those
interactions are still somewhat unclear this will have to do. See #1352
for plans in that direction.

Fixes #511.