Fix customized built-in element is handling and reactions
This uses the new "is value" concept introduced to DOM in
https://github.com/whatwg/dom/pull/262, which more correctly implements
the intention that changing the is attribute after element creation does
not impact the processing model, and thus fixes #1298. This allows us to
then fix #1360: custom element callback reactions for customized
built-in elements were not correctly looking up their definition.