Enqueue custom element reactions appropriately during upgrades
As pointed out in
https://github.com/whatwg/html/issues/1297#issuecomment-222890386, the
setup introduced in #1309 allows custom element reactions to happen
during custom element constructors. This is unavoidable, but the spec
did not account for its consequences, leading to cases where a custom
element reaction would be called twice for the same attribute (or the
same element being connected). By moving these steps to the top of the
algorithm, we ensure that reactions are only enqueued for the state of
the world before the constructor runs, and that any reactions caused by
actions inside the constructor are then taken care of by the usual
mechanisms.

Closes #1297 again.