Refine some autcomplete attribute examples
autocomplete=street-address isn’t allowed with input type=text, because
the street-address field name is for a multiline street address. So,
change the example to use texarea, which does allow
autocomplete=street-address.

Also, add a note about the <input type=month autocomplete="bday">
example being non-conforming (which it’s OK for it to be, since the
section it’s in provides UA requirements and so is targeted to
implementors, and not intended for authors).