Take care of missing custom element reactions
As discussed in https://github.com/w3c/webcomponents/issues/186, there
are certain scenarios where custom element reactions could be enqueued,
but the custom element reactions stack is empty. The most prominent of
these is editing operations.

We solve this issue in two parts:

- First, introduce the backup element queue (similar to the old "base
element queue" concept; see
https://github.com/w3c/webcomponents/issues/457). This takes care of
cases like editing which cannot be solved in any other way.
- Also, we mandate that [CEReactions] be applied to nonstandard APIs
that modify the DOM, including a list of what we've identified so far.

Closes https://github.com/w3c/webcomponents/issues/186.