Disallow mismatches between custom element local names and brands
Without this fix, it is possible to install the brand for one type of
element on an element with a different local name, in one of two
different ways (both shown here as examples). This fix makes the super()
call throw when this is attempted, preserving the invariant that a brand
is only installed on an element with the correct local name.

Originally discussed at
https://bugs.chromium.org/p/chromium/issues/detail?id=619062.