Prevent double custom element reaction enqueuing via upgrades
As already discussed in the spec (as of #1365), it is sometimes possible
to enqueue two upgrade reactions, via a constructor that does things you
are not supposed to do. The fix in #1365 prevented upgrading twice.
However, it did not bail out early enough, and allowed
attributeChangedCallback and connectedCallback reactions to be enqueued
(redundantly) before the bail-out. This commit moves the bail-out logic
to the beginning of the algorithm, so as to avoid any such
double-enqueuing.

Fixes https://github.com/w3c/webcomponents/issues/523.