Clarify the memoization performed by the module map
The previous statement, "once per Document", was slightly incorrect,
since module maps are actually associated with environment settings
objects. Environment settings objects are 1:1 with global objects (such
as Windows and WorkerGlobalScopes), but Windows are not 1:1 with
Documents (due to document.open() and the initial about:blank).

The new phrasing retains the readability of the old one, by using
"Window or WorkerGlobalScope" instead of just saying "environment
settings object", but it is more accurate.