Fix #1099 Revert behavior of 2d context "currentColor" keyword.
This change reverts a possibly unintentional change in behavior
in commit 285b52f that made "currentColor" evaluate to opaque
black for canvases that are not rendered.  Going back to the
old behavior, which reflect most current implementations. The old
behavior is to make "currentColor" evaluate to opaque black only
when the computed "color" style of the associated canvas is undefined.