Revert behavior change of 2D context "currentColor" keyword
This change reverts a possibly unintentional change in behavior in
commit 285b52f that made "currentColor" evaluate to opaque black for
canvases that are not rendered. This goes back to the old behavior,
which reflects most current implementations. The old behavior is to make
"currentColor" evaluate to opaque black only when the computed "color"
style of the associated canvas is undefined.

Fixes #1099.