Change minlength/maxlength behavior around linebreaks and code points
As discussed in #1467, the current situation around these attributes is
not very interoperable in two ways:

* Some browsers count line breaks as two characters (per the spec),
others as one character.
* Some browsers count code units (per the spec), others code points,
others grapheme clusters.

Per discussions, this updates minlength and maxlength to count line
breaks as one character and to count code points. We believe this is the
most developer-friendly approach. Notably the choice of code points
aligns with the pattern="" attribute.

While here, updated the textLength property to return the code-unit
length of the element's API value, instead of the element's value, since
as per http://software.hixie.ch/utilities/js/live-dom-viewer/?saved=4304
that is what browsers actually do.

Fixes #1467.