Be more precise about nested browsing contexts
This commit tries to clear up some of the confusion in the current spec
around nested browsing contexts and how they impact the processing
model.

First, many of the definitions in the "nested browsing context" section
are rephrased for clarity. Per #1180, the term "nested browsing context"
itself becomes a field of a "browsing context container". The field has
values of either a browsing context, or null. Then, all algorithms which
refer to a browsing context container's nested browsing context are
explicitly updated to account for the null case. In particular, this
makes it clearer that when a browsing context is discarded, the
associated browsing context container has its nested browsing context
set to null.

Then, we make a normative change to make window.{top, parent,
frameElement} all return null when their associated browsing context has
been discarded. As discussed in #1279, browser behavior here varies
extensively, but nobody seems to follow the current spec at all, so we
went with the 2/4 majority of returning null.

Along the way, this clarifies some of the garbage collection behavior,
and adds examples for top/parent/frameElement.

Fixes #1180. Fixes #1279.