Prevent double custom element constructor calls
If creating an element failed the first time due a misbehaving
constructor, it was possible that in the future we would try to upgrade
the element, running the constructor again. This commit uses the
newly-introduced (as of https://github.com/whatwg/dom/pull/282) custom
element state of "failed" to prevent upgrading such elements.

Fixes https://github.com/w3c/webcomponents/issues/533.